Added a buffered log message system
Instead of storing log messages in the console directly, they are now
stored inside of an internal log message buffer, which is then used to
render the text in the console. This allows for proper filtering of log
levels without losing information. Also opens up the possibility of
writing to log files.