Created LambdaEventFilter template class.
- Created LambdaEventFilter, which acts like a regular event filter that
can be installed onto objects using installEventFilter, but a lambda
function as well as arguments can be passed to the constructor, which
will LambdaEventFilter's implementation of the eventFilter virtual
function will execute. Effectively, this allows the event filter to
execute any sort of code, removing the need to create multiple event
filter classes and overriding the eventFilter virtual function multiple
times, instead a new instance of LambdaEventFilter can be constructed
but given a different lambda, making it behave differently.

- Stylized signal/slot syntax a little.