Added process image scanner mode support
The WindowTreeDialog has been transformed into ProcessScannerDialog, and
now takes a ProcessScanner::SCAN_SCOPE as a constructor argument, which
determines the mode of the dialog (Window vs Process). Added another
select button for the process image mode which spawns the process
scanner in process mode, allowing for the selection of processes as well
as windows. Renamed relevant files and variables to reflect changes.